+++ object_database/web/content/src/components/ImageFigure.js
@@ -38,7 +38,7 @@ class ImageFigure {
 
         this.triangleBuffer = gl.createBuffer();
         gl.bindBuffer(gl.ARRAY_BUFFER, this.triangleBuffer);
-        gl.bufferData(gl.ARRAY_BUFFER, new Float32Array([0.0, 0.0, 0.0, 1.0, 1.0, 0.0, 1.0, 1.0]), gl.STATIC_DRAW);
+        gl.bufferData(gl.ARRAY_BUFFER, new Float64Array([0.0, 0.0, 0.0, 1.0, 1.0, 0.0, 1.0, 1.0]), gl.STATIC_DRAW);
     }
 
     clear(renderer) {
